Sat Update
Lots of bug fixes / improvement to performance

Removed more redundant code

Fixed the close SCT button ( wasnt closing all instances of SCTools)

fixed SCQC Button scaling to update when window scale is changed

A large part of this tool is essentially macros( move mouse here, click, type, move mouse back)
If the location that the mouse clicks during operation is obscured it causes the operation to fail.
I have set the location that the mouse moves to be to the left of the screen.
And now should be more reliable.

I